+++ INTERX/tasks/node_discover.go
@@ -42,8 +42,12 @@ func timeout() time.Duration {
 	return timeoutDuration
 }
 
-func getRPCAddress(ipAddr string) string {
+func getTendermintRPCAddress(ipAddr string) string {
 	return "http://" + ipAddr + ":" + config.Config.NodeDiscovery.DefaultTendermintPort
+}
+
+func getInterxAddress(ipAddr string) string {
+	return "http://" + ipAddr + ":" + config.Config.NodeDiscovery.DefaultInterxPort
 }
 
 func QueryPeers(rpcAddr string) ([]tmTypes.Peer, error) {
@@ -160,7 +164,7 @@ func NodeDiscover(rpcAddr string, isLog bool) {
 				common.GetLogger().Info("[node-discovery] ", ipAddr)
 			}
 
-			kiraStatus, err := QueryKiraStatus(getRPCAddress(ipAddr))
+			kiraStatus, err := QueryKiraStatus(getTendermintRPCAddress(ipAddr))
 			if err != nil {
 				continue
 			}
@@ -184,7 +188,7 @@ func NodeDiscover(rpcAddr string, isLog bool) {
 			}
 
 			if _, ok := peersFromIP[ipAddr]; !ok {
-				peersFromIP[ipAddr], _ = QueryPeers(getRPCAddress(ipAddr))
+				peersFromIP[ipAddr], _ = QueryPeers(getTendermintRPCAddress(ipAddr))
 			}
 
 			peers := peersFromIP[ipAddr]
@@ -216,6 +220,23 @@ func NodeDiscover(rpcAddr string, isLog bool) {
 			}
 
 			PubP2PNodeListResponse.NodeList = append(PubP2PNodeListResponse.NodeList, nodeInfo)
+
+			// interxStatus := common.GetInterxStatus(getInterxAddress(ipAddr))
+			interxStatus := common.GetInterxStatus(getInterxAddress("127.0.0.1"))
+			common.GetLogger().Info(interxStatus)
+
+			if interxStatus != nil {
+				interxInfo := types.InterxNode{}
+				interxInfo.ID = interxStatus.ID
+				interxInfo.IP = ipAddr
+				interxInfo.Ping = 0 // calculate and verify
+				interxInfo.Moniker = interxStatus.InterxInfo.Moniker
+				interxInfo.Faucet = interxStatus.InterxInfo.FaucetAddr
+				interxInfo.Type = interxStatus.InterxInfo.Node.NodeType
+				interxInfo.Version = interxStatus.InterxInfo.Version
+
+				InterxP2PNodeListResponse.NodeList = append(InterxP2PNodeListResponse.NodeList, interxInfo)
+			}
 		}
 
 		lastUpdate := time.Now().Unix()